refactoring condition of when to delete ckpt files when saving a new ckpt file (to not delete files used for inference eval in train loop)

+++ scripts/training/prod/train.py
@@ -559,7 +559,7 @@ def save_ckpt(
 
     os.makedirs(f"{ckpt_dir}/{exp_name}_{run_id}", exist_ok=True)
 
-    if file_name != "latesttrain":
+    if file_name != "latesttrain" and "inf" not in file_name:
         if len(glob.glob(f"{ckpt_dir}/{exp_name}_{run_id}/{file_name}_*.pt")) > 0:
             for p in glob.glob(f"{ckpt_dir}/{exp_name}_{run_id}/{file_name}_*.pt"):
                 os.remove(p)
